Remove Pending Tab
